Unbundle custom events code
